refactor(Number): Number 패키지 수정
