+++ src/main/java/step3and4and5/client/number/RandomNumber.java
@@ -1,5 +1,7 @@
 package step3and4and5.client.number;
 
+import step3and4and5.game.car.Number;
+
 import java.util.concurrent.ThreadLocalRandom;
 
 public final class RandomNumber implements Number {
