+++ src/main/java/step3and4and5/game/CarGame.java
@@ -1,9 +1,9 @@
 package step3and4and5.game;
 
 import step3and4and5.client.input.InputView;
-import step3and4and5.client.number.Number;
 import step3and4and5.client.output.OutputView;
 import step3and4and5.game.car.Cars;
+import step3and4and5.game.car.Number;
 import step3and4and5.game.car.WinningCars;
 
 import java.io.IOException;
