+++ src/main/java/step3and4and5/game/car/Cars.java
@@ -1,7 +1,5 @@
 package step3and4and5.game.car;
 
-import step3and4and5.client.number.Number;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
