+++ src/test/java/step3and4and5/game/car/CarTest.java
@@ -2,7 +2,6 @@ package step3and4and5.game.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3and4and5.client.number.Number;
 
 import static org.assertj.core.api.Assertions.assertThat;
