+++ src/test/java/step3and4and5/game/car/CarsTest.java
@@ -2,7 +2,6 @@ package step3and4and5.game.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3and4and5.client.number.Number;
 import step3and4and5.client.number.RandomNumber;
 
 import java.util.List;
@@ -49,4 +48,5 @@ class CarsTest {
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessage("빈 공백일 수 없습니다.");
     }
+
 }
